+++ src/main/java/org/example/parser/impl/NumTurnsParser.java
@@ -5,11 +5,11 @@ import org.example.parser.Parser;
 import java.io.File;
 import java.io.IOException;
 
-public class NumTurnsParser extends Parser<Integer> {
+public class NumTurnsParser extends Parser<Long> {
     private static final String NUM_TURNS_KEY = "num_turns";
 
     @Override
-    public Integer parse(File file) throws IOException {
-        return parseJson(file).get(NUM_TURNS_KEY).asInt();
+    public Long parse(File file) throws IOException {
+        return parseJson(file).get(NUM_TURNS_KEY).asLong();
     }
 }
